+++ runtime-jvm/build.sbt
@@ -39,33 +39,13 @@ lazy val root = project.in(file(".")).
   settings(commonSettings).
   aggregate(main, benchmark)
 
-lazy val main = project.in(file("main")).
-  settings(commonSettings).
-  settings(name := "unison-runtime").
-  settings(sourceGenerators in Compile += Def.task {
-    import org.unisonweb.codegeneration._
-    val outPath = (sourceManaged in Compile).value / "org" / "unisonweb" / "compilation"
-    val gens : List[(File, String)] = List[OneFileGenerator](
-      ComputationGenerator,
-      ValueGenerator,
-      CompileLookupVarGenerator,
-      CompileLambdaGenerator,
-      CompileLet1Generator,
-      CompileLetRecGenerator,
-      CompileFunctionApplicationGenerator,
-//      MakeBuiltinGenerator,
-      TailCallsGenerator,
-      CompileIf0Generator
-    ).map(_.apply(outPath))
+lazy val main = project.in(file("main"))
+  .settings(commonSettings)
+  .settings(name := "unison-runtime")
 
-    gens.map { case (file, content) => IO.write(file, content); file: File }
-  }.taskValue)
-
-lazy val benchmark = project.in(file("benchmark")).
-  settings(commonSettings).
-  settings(
-    name := "unison-runtime-benchmark"
-  )
+lazy val benchmark = project.in(file("benchmark"))
+  .settings(commonSettings)
+  .settings(name := "unison-runtime-benchmark")
   .settings(
     libraryDependencies += scalaOrganization.value % "scala-reflect" % scalaVersion.value
   )
